refactor(example): modularize production-api into clean architecture
