+++ examples/production-api/src/application/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod notes;
