+++ examples/production-api/src/domain/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod health;
+pub(crate) mod notes;
