+++ examples/production-api/src/presentation/handlers/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod drills;
+pub(crate) mod health;
+pub(crate) mod notes;
