+++ examples/production-api/src/tests/mod.rs
@@ -1,0 +1,4 @@
+mod config;
+mod pagination;
+mod router;
+mod testkit;
